fix: publish release
affects: @overmindbots/bot-referral-ranks-fulfillment-service

+++ scripts/deployDev.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 set -e
 
+# Cheap arg handing
+argOne=$1
+
 devDeployTimestamp=$(date +%s)
 monorepoImageUrl="gcr.io/overmindbots/core:$devDeployTimestamp"
 
@@ -10,13 +13,18 @@ rm -rf k8s-generated
 mkdir k8s-generated
 touch k8s-generated/.gitkeep
 
-{
-  rm .devDeploy
-} || {
-  echo "Creating .devDeploy file"
-}
-touch .devDeploy
-echo $devDeployTimestamp > .devDeploy
+if [[ $argOne == '--no-build' ]] ; then
+  echo "!!!! No build flag used! Only deploying to local kubernetes !!!!"
+else
+  # We update the .devDeploy version
+  {
+    rm .devDeploy
+  } || {
+    echo "Creating .devDeploy file"
+  }
+  touch .devDeploy
+  echo $devDeployTimestamp > .devDeploy
+fi
 
 {
   helm install --name overmindbots-mongodb stable/mongodb --namespace overmindbots --set mongodbUsername=dev,mongodbPassword=dev,mongodbDatabase=overmindbots
@@ -31,8 +39,11 @@ echo $devDeployTimestamp > .devDeploy
 }
 
 # === Build image === #
-
-docker build --no-cache -t $monorepoImageUrl -f Dockerfile .
+if [[ $argOne == '--no-build' ]] ; then
+  echo "=> Skipping docker build"
+else
+  docker build --no-cache -t $monorepoImageUrl -f Dockerfile .
+fi
 
 node ./scripts/buildTemplates.js service-referral-ranks-invites
 node ./scripts/buildTemplates.js service-bot-manager
